Move Procedure and BiProcedure to procedure package.
